Refactoring user model
